+++ SistemaPOS/CapaDatos/Entity/Cliente.cs
@@ -21,11 +21,11 @@ namespace CapaDatos.Entity
         }
     
         public int idCliente { get; set; }
-        public Nullable<int> dni { get; set; }
+        public Nullable<long> dni { get; set; }
         public string apellido { get; set; }
         public string nombre { get; set; }
         public string email { get; set; }
-        public Nullable<int> telefono { get; set; }
+        public Nullable<long> telefono { get; set; }
         public string direccion { get; set; }
         public Nullable<int> estado { get; set; }
         public Nullable<System.DateTime> fechaAlta { get; set; }
